+++ tap_s3_csv/sync.py
@@ -44,6 +44,9 @@ def sync_stream(config, state, table_spec, stream):
         state = singer.write_bookmark(state, table_name, 'modified_since', s3_file['last_modified'].isoformat())
         singer.write_state(state)
 
+    if s3.skipped_files_count:
+        LOGGER.warn("%s files got skipped during the last sync.",s3.skipped_files_count)
+
     LOGGER.info('Wrote %s records for table "%s".', records_streamed, table_name)
 
     return records_streamed
@@ -56,13 +59,20 @@ def sync_table_file(config, s3_path, table_spec, stream):
     # Check whether file is without extension or not
     if not extension or s3_path.lower() == extension:
         LOGGER.warning('"%s" without extension will not be synced.',s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
         return 0
-    if extension == "zip":
-        return sync_compressed_file(config, s3_path, table_spec, stream)
-    if extension in ["csv", "gz", "jsonl", "txt"]:
-        return handle_file(config, s3_path, table_spec, stream, extension)
-
-    LOGGER.warning('"%s" having the ".%s" extension will not be synced.',s3_path,extension)
+    try:
+        if extension == "zip":
+            return sync_compressed_file(config, s3_path, table_spec, stream)
+        if extension in ["csv", "gz", "jsonl", "txt"]:
+            return handle_file(config, s3_path, table_spec, stream, extension)
+        LOGGER.warning('"%s" having the ".%s" extension will not be synced.',s3_path,extension)
+    except (UnicodeDecodeError,json.decoder.JSONDecodeError):
+        # UnicodeDecodeError will be raised if non csv file passed to csv parser
+        # JSONDecodeError will be raised if non JSONL file passed to JSON parser
+        # Handled both error and skipping file with wrong extension.
+        LOGGER.warning("Skipping %s file as parsing failed. Verify an extension of the file.",s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
     return 0
 
 
@@ -75,6 +85,7 @@ def handle_file(config, s3_path, table_spec, stream, extension, file_handler = N
     # Check whether file is without extension or not
     if not extension or s3_path.lower() == extension:
         LOGGER.warning('"%s" without extension will not be synced.',s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
         return 0
     if extension == "gz":
         return sync_gz_file(config, s3_path, table_spec, stream, file_handler)
@@ -89,19 +100,27 @@ def handle_file(config, s3_path, table_spec, stream, extension, file_handler = N
 
         # If file is extracted from zip or gz use file object else get file object from s3 bucket
         file_handle = file_handler if file_handler else s3.get_file_handle(config, s3_path)._raw_stream
-        return sync_jsonl_file(config, file_handle, s3_path, table_spec, stream)
+        records =  sync_jsonl_file(config, file_handle, s3_path, table_spec, stream)
+        if records == 0:
+            # Only space isn't the valid JSON but it is a valid CSV header hence skipping the jsonl file with only space.
+            s3.skipped_files_count = s3.skipped_files_count + 1
+            LOGGER.warning('Skipping "%s" file as it is empty', s3_path)
+        return records
 
     if extension == "zip":
         LOGGER.warning('Skipping "%s" file as it contains nested compression.',s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
         return 0
 
     LOGGER.warning('"%s" having the ".%s" extension will not be synced.',s3_path,extension)
+    s3.skipped_files_count = s3.skipped_files_count + 1
     return 0
 
 
 def sync_gz_file(config, s3_path, table_spec, stream, file_handler):
     if s3_path.endswith(".tar.gz"):
         LOGGER.warning('Skipping "%s" file as .tar.gz extension is not supported',s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
         return 0
 
     # If file is extracted from zip use file object else get file object from s3 bucket
@@ -110,12 +129,22 @@ def sync_gz_file(config, s3_path, table_spec, stream, file_handler):
     file_bytes = file_object.read()
     gz_file_obj = gzip.GzipFile(fileobj=io.BytesIO(file_bytes))
 
-    gz_file_name = utils.get_file_name_from_gzfile(fileobj=io.BytesIO(file_bytes))
+    # pylint: disable=duplicate-code
+    try:
+        gz_file_name = utils.get_file_name_from_gzfile(fileobj=io.BytesIO(file_bytes))
+    except AttributeError as err:
+        # If a file is compressed using gzip command with --no-name attribute,
+        # It will not return the file name and timestamp. Hence we will skip such files.
+        # We also seen this issue occur when tar is used to compress the file
+        LOGGER.warning('Skipping "%s" file as we did not get the original file name',s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
+        return 0
 
     if gz_file_name:
 
         if gz_file_name.endswith(".gz"):
             LOGGER.warning('Skipping "%s" file as it contains nested compression.',s3_path)
+            s3.skipped_files_count = s3.skipped_files_count + 1
             return 0
 
         gz_file_extension = gz_file_name.split(".")[-1].lower()
@@ -159,26 +188,38 @@ def sync_csv_file(config, file_handle, s3_path, table_spec, stream):
     # memory consumption but that's acceptable as well.
     csv.field_size_limit(sys.maxsize)
 
-    iterator = csv_helper.get_row_iterator(
-        file_handle, table_spec, stream["schema"]["properties"].keys(), True)
+    if "properties" in stream["schema"]:
+        iterator = csv_helper.get_row_iterator(
+            file_handle, table_spec, stream["schema"]["properties"].keys(), True)
+    else:
+        iterator = csv_helper.get_row_iterator(file_handle, table_spec, None, True)
 
     records_synced = 0
 
-    for row in iterator:
-        custom_columns = {
-            s3.SDC_SOURCE_BUCKET_COLUMN: bucket,
-            s3.SDC_SOURCE_FILE_COLUMN: s3_path,
+    if iterator:
+        for row in iterator:
 
-            # index zero, +1 for header row
-            s3.SDC_SOURCE_LINENO_COLUMN: records_synced + 2
-        }
-        rec = {**row, **custom_columns}
+            #Skipping the empty line of CSV
+            if len(row) == 0:
+                continue
 
-        with Transformer() as transformer:
-            to_write = transformer.transform(rec, stream['schema'], metadata.to_map(stream['metadata']))
+            custom_columns = {
+                s3.SDC_SOURCE_BUCKET_COLUMN: bucket,
+                s3.SDC_SOURCE_FILE_COLUMN: s3_path,
 
-        singer.write_record(table_name, to_write)
-        records_synced += 1
+                # index zero, +1 for header row
+                s3.SDC_SOURCE_LINENO_COLUMN: records_synced + 2
+            }
+            rec = {**row, **custom_columns}
+
+            with Transformer() as transformer:
+                to_write = transformer.transform(rec, stream['schema'], metadata.to_map(stream['metadata']))
+
+            singer.write_record(table_name, to_write)
+            records_synced += 1
+    else:
+        LOGGER.warning('Skipping "%s" file as it is empty',s3_path)
+        s3.skipped_files_count = s3.skipped_files_count + 1
 
     return records_synced
 
@@ -196,6 +237,9 @@ def sync_jsonl_file(config, iterator, s3_path, table_spec, stream):
         decoded_row = row.decode('utf-8')
         if decoded_row.strip():
             row = json.loads(decoded_row)
+            # Skipping the empty json row.
+            if len(row) == 0:
+                continue
         else:
             continue