+++ tests/test_all_supported_files.py
@@ -1,26 +1,27 @@
 import simplejson
-import tap_tester.connections as connections
-import tap_tester.menagerie   as menagerie
-import tap_tester.runner      as runner
+from tap_tester import connections, menagerie, runner
+from base import S3CSVBaseTest
 
 import utils_for_test as utils
 import unittest
 
 ALL_SUPPORTED_FOLDER_PATH = "All-Supported-Files"
 
-class S3AllFilesSupport(unittest.TestCase):
+class S3AllFilesSupport(S3CSVBaseTest):
+
+    table_entry = [
+        {'table_name': 'all_support_csv', 'search_prefix': 'int_test_all_support_csv', 'search_pattern': 'int_test_all_support_csv\\/.*\\.csv'},
+        {'table_name': 'all_support_jsonl', 'search_prefix': 'int_test_all_support_jsonl', 'search_pattern': 'int_test_all_support_jsonl\\/.*\\.jsonl'},
+        {'table_name': 'all_support_gz_has_csv', 'search_prefix': 'int_test_all_support_gz_has_csv', 'search_pattern': 'int_test_all_support_gz_has_csv\\/.*\\.gz'},
+        {'table_name': 'all_support_gz_has_jsonl', 'search_prefix': 'int_test_all_support_gz_has_jsonl', 'search_pattern': 'int_test_all_support_gz_has_jsonl\\/.*\\.gz'},
+        {'table_name': 'all_support_zip', 'search_prefix': 'int_test_all_support_zip', 'search_pattern': 'int_test_all_support_zip\\/.*\\.zip'}
+    ]
 
     def setUp(self):
         self.conn_id = connections.ensure_connection(self)
 
     def resource_names(self):
         return ["sample_csv_file.csv","sample_jsonl_file.jsonl","sample_gz_file_has_csv.gz","sample_gz_file_has_jsonl.gz","sample_zip_file.zip"]
-
-    def tap_name(self):
-        return "tap-s3-csv"
-
-    def get_type(self):
-        return "platform.s3-csv"
 
     def name(self):
         return "test_all_supported_files"
@@ -44,33 +45,13 @@ class S3AllFilesSupport(unittest.TestCase):
         }
 
     def expected_pks(self):
-        return {}
-
-    def get_credentials(self):
-        return {}
-
-    def get_properties(self):
         return {
-            'start_date' : '2017-01-01T00:00:00Z',
-            "bucket": "com-stitchdata-prod-circleci-assets",
-            "account_id": "218546966473",
-            'tables' : "[{\"table_name\": \"all_support_csv\",\"search_prefix\": \"int_test_all_support_csv\",\"search_pattern\": \"int_test_all_support_csv\\\\/.*\\\\.csv\"},{\"table_name\": \"all_support_jsonl\",\"search_prefix\": \"int_test_all_support_jsonl\",\"search_pattern\": \"int_test_all_support_jsonl\\\\/.*\\\\.jsonl\"},{\"table_name\": \"all_support_gz_has_csv\",\"search_prefix\": \"int_test_all_support_gz_has_csv\",\"search_pattern\": \"int_test_all_support_gz_has_csv\\\\/.*\\\\.gz\"},{\"table_name\":\"all_support_gz_has_jsonl\",\"search_prefix\": \"int_test_all_support_gz_has_jsonl\",\"search_pattern\": \"int_test_all_support_gz_has_jsonl\\\\/.*\\\\.gz\"},{\"table_name\": \"all_support_zip\",\"search_prefix\": \"int_test_all_support_zip\",\"search_pattern\": \"int_test_all_support_zip\\\\/.*\\\\.zip\"}]"
+            "all_support_csv": {},
+            "all_support_jsonl": {},
+            "all_support_gz_has_csv": {},
+            "all_support_gz_has_jsonl": {},
+            "all_support_zip": {}
         }
-
-
-    def select_found_catalogs(self, found_catalogs):
-        # selected = [menagerie.select_catalog(self.conn_id, c) for c in found_catalogs]
-        # menagerie.post_annotated_catalogs(self.conn_id, selected)
-        for catalog in found_catalogs:
-            schema = menagerie.get_annotated_schema(self.conn_id, catalog['stream_id'])
-            non_selected_properties = []
-            additional_md = []
-
-            connections.select_catalog_and_fields_via_metadata(
-                self.conn_id, catalog, schema, additional_md=additional_md,
-                non_selected_fields=non_selected_properties
-            )
-
 
     def setUpTestEnvironment(self):
         index = 0
@@ -78,26 +59,21 @@ class S3AllFilesSupport(unittest.TestCase):
             utils.delete_and_push_file(self.get_properties(), [resource], ALL_SUPPORTED_FOLDER_PATH, index)
             index += 1
 
-
     def test_run(self):
 
-        self.setUpTestEnvironment()
+        self.setUpCompressedEnv(ALL_SUPPORTED_FOLDER_PATH)
 
-        runner.run_check_job_and_check_status(self)
-
-        found_catalogs = menagerie.get_catalogs(self.conn_id)
-        self.assertEqual(len(found_catalogs), len(self.expected_check_streams()), msg="unable to locate schemas for connection {}".format(self.conn_id))
-
-        found_catalog_names = set(map(lambda c: c['tap_stream_id'], found_catalogs))
-        subset = self.expected_check_streams().issubset( found_catalog_names )
-        self.assertTrue(subset, msg="Expected check streams are not subset of discovered catalog")
+        found_catalogs = self.run_and_verify_check_mode(self.conn_id)
 
         # Clear state before our run
         menagerie.set_state(self.conn_id, {})
 
-        self.select_found_catalogs(found_catalogs)
+        # Select our catalogs
+        our_catalogs = [c for c in found_catalogs if c.get('tap_stream_id') in self.expected_sync_streams()]
 
-        runner.run_sync_job_and_check_status(self)
+        self.perform_and_verify_table_and_field_selection(self.conn_id, our_catalogs, True)
+
+        self.run_and_verify_sync(self.conn_id)
 
         no_csv_records = 998
         no_jsonl_records = 10