+++ CHANGELOG.md
@@ -3,6 +3,32 @@
 ## 1.4.0
   * Updates for python 3.11 [#61](https://github.com/singer-io/tap-s3-csv/pull/61)
 
+## 1.3.7
+  * Remove Backoff for Access Denied errors
+  * Add unitttest for access denied giveup
+  * [#55](https://github.com/singer-io/tap-s3-csv/pull/55)
+
+## 1.3.6
+  *  Implement request timeout [#46](https://github.com/singer-io/tap-s3-csv/pull/46)
+  *  Updgrade SDK version [#51](https://github.com/singer-io/tap-s3-csv/pull/51)
+  *  Add integration tests [#52](https://github.com/singer-io/tap-s3-csv/pull/52)
+
+## 1.3.5
+  * Increase the limit on the width of a field in the CSV files read by the tap [#47](https://github.com/singer-io/tap-s3-csv/pull/47)
+
+## 1.3.4
+  * Bump singer-encodings to 0.1.2 [#45](https://github.com/singer-io/tap-s3-csv/pull/45)
+
+## 1.3.3
+  * Bump singer-encodings to 0.1.1 [#41](https://github.com/singer-io/tap-s3-csv/pull/41)
+
+## 1.3.2
+  * Skip files without a name [#37](https://github.com/singer-io/tap-s3-csv/pull/37)
+  * Fix an issue to allow the tap to run with a catalog without schemas [#38](https://github.com/singer-io/tap-s3-csv/pull/38)
+
+## 1.3.1
+  * Fixed bug that caused `integer`s to be discovered as `number` differently in different versions of python [#35](https://github.com/singer-io/tap-s3-csv/pull/35)
+
 ## 1.3.0
   * Adds support for Compressed files [#32](https://github.com/singer-io/tap-s3-csv/pull/32)
   * Adds support for JSONL files [#31](https://github.com/singer-io/tap-s3-csv/pull/31)